FIX: Revenues were not calculated correctly in some cases.

+++ gui/goods_stats_t.cc
@@ -98,8 +98,8 @@ void goods_stats_t::zeichnen(koord offset)
 			}
 			else
 			{
-				const uint8 differential = journey_minutes -welt->get_settings().get_tolerable_comfort_short_minutes();
-				const uint8 max_differential = welt->get_settings().get_tolerable_comfort_median_long_minutes() -welt->get_settings().get_tolerable_comfort_short_minutes();
+				const uint16 differential = journey_minutes - welt->get_settings().get_tolerable_comfort_short_minutes();
+				const uint16 max_differential = welt->get_settings().get_tolerable_comfort_median_long_minutes() -welt->get_settings().get_tolerable_comfort_short_minutes();
 				const uint32 proportion = differential * 100 / max_differential;
 				comfort_modifier = (80ll * (sint64)proportion / 100ll) + 20ll;
 			}
